Replace the old `ContextMenu` file with the new one
